requirements.txt
torch>=2.0.0
diffusers>=0.24.0
Pillow>=10.0.0
customtkinter>=5.2.0